api/requirements.txt
Pillow>=10.3.0
lxml>=5.2.1
pydantic>=2.7.0
